Refactored code, so the client is only created once. This avoids multiple connections.
