+++ controller/mqtt/mqttSubscribe.js
@@ -1,19 +1,8 @@
 require('dotenv').config();
 
-const mqtt = require('async-mqtt');
-
 module.exports = mqttSubscribe = async (client, subTopic) => {
   return new Promise(async (resolve, reject) => {
     try {
-      // TODO: Handle server crash when authentication fails
-      console.log('Attempting MQTT connection with', process.env.MQTT_URL);
-      const client = mqtt.connect(
-        `${process.env.MQTT_URL}:${process.env.MQTT_PORT}`,
-        {
-          username: process.env.MQTT_USER,
-          password: process.env.MQTT_PASSWORD,
-        }
-      );
       client.on('connect', async () => {
         await client.subscribe(subTopic);
         console.log('Subscribed to topic', subTopic);
@@ -21,7 +10,6 @@ module.exports = mqttSubscribe = async (client, subTopic) => {
       // TODO: Timeout
       client.on('message', async (topic, payload) => {
         // TODO: BugFix: we should look for a specific topic
-        await client.end();
         console.log('Message received');
         resolve({ topic: topic, message: payload.toString() });
       });
