+++ controller/powerDevice.js
@@ -1,5 +1,7 @@
 require('dotenv').config();
 
+const mqtt = require('async-mqtt');
+
 const mqttPost = require('./mqtt/mqttPost');
 const mqttSubscribe = require('./mqtt/mqttSubscribe');
 
@@ -8,8 +10,21 @@ module.exports = powerDevice = async (_id, serialNumber, command) => {
     `${command === '2' ? `Toggling` : `Powering ${command}`} device`,
     serialNumber
   );
-  await mqttPost(`cmnd/${serialNumber}_fb/power`, command);
-  const mqttResponse = await mqttSubscribe(`stat/${serialNumber}_fb/POWER`);
+  // TODO: Handle server crash when authentication fails
+  console.log('Attempting MQTT connection with', process.env.MQTT_URL);
+  const client = mqtt.connect(
+    `${process.env.MQTT_URL}:${process.env.MQTT_PORT}`,
+    {
+      username: process.env.MQTT_USER,
+      password: process.env.MQTT_PASSWORD,
+    }
+  );
+  await mqttPost(client, `cmnd/${serialNumber}_fb/power`, command);
+  const mqttResponse = await mqttSubscribe(
+    client,
+    `stat/${serialNumber}_fb/POWER`
+  );
+  await client.end();
   // The response expected by the front end should be something like this:
   // {
   //   "_id": "XXXXXXXXXX",
